Save Search modal now saves user defined Search
- moved props, state, and api call from Dashboard to new Saved Search Modal.
- still need to clean up code.
- still need to round out create vuln functionality.
